+++ test/Makefile.am
@@ -13,7 +13,27 @@ CXXFLAGS += -Wall --std=c++11 -g $(PTHREAD_CFLAGS)
 TESTS = arestest
 
 noinst_PROGRAMS = arestest aresfuzz
-arestest_SOURCES = ares-test-main.cc ares-test-init.cc ares-test.cc ares-test-parse.cc ares-test-misc.cc ares-test-live.cc ares-test-mock.cc ares-test-internal.cc dns-proto.cc dns-proto-test.cc ares-test.h dns-proto.h
+arestest_SOURCES = ares-test-main.cc \
+	ares-test-init.cc \
+	ares-test.cc \
+	ares-test-parse.cc \
+	ares-test-parse-a.cc \
+	ares-test-parse-aaaa.cc \
+	ares-test-parse-mx.cc \
+	ares-test-parse-naptr.cc \
+	ares-test-parse-ns.cc \
+	ares-test-parse-ptr.cc \
+	ares-test-parse-soa.cc \
+	ares-test-parse-srv.cc \
+	ares-test-parse-txt.cc \
+	ares-test-misc.cc \
+	ares-test-live.cc \
+	ares-test-mock.cc \
+	ares-test-internal.cc \
+	dns-proto.cc \
+	dns-proto-test.cc \
+	ares-test.h \
+	dns-proto.h
 arestest_LDADD = libgmock.la libgtest.la $(ARES_BLD_DIR)/libcares.la $(PTHREAD_LIBS)
 
 # Not interested in coverage of test code, but linking the test binary needs the coverage option
